Change MassUnits from float to double for compatibility.
--HG--
branch : week-of-code

+++ src/enzo/gFLDProblem.h
@@ -193,7 +193,7 @@ class gFLDProblem : public virtual NonlinearProblemABC {
   float TimeUnits;     // time scaling factor
   float TempUnits;     // temperature scaling factor
   float VelUnits;      // velocity scaling factor
-  float MassUnits;     // mass scaling factor
+  double MassUnits;    // mass scaling factor
 
   // chemistry constants
   float HFrac;         // Fraction of matter composed of Hydrogen
